refactor: endpointProperties 분리
